+++ src/game/rotation.c
@@ -17,8 +17,8 @@ void	rotate_left(t_player *p)
 	p->angle -= ROTATE_AMOUNT;
 	if (p->angle < 0)
 		p->angle += PI * 2.0;
-	p->dir_x = cos(p->angle);
-	p->dir_y = sin(p->angle);
+	p->dir_x = cosf(p->angle);
+	p->dir_y = sinf(p->angle);
 }
 
 void	rotate_right(t_player *p)
@@ -26,6 +26,6 @@ void	rotate_right(t_player *p)
 	p->angle += ROTATE_AMOUNT;
 	if (p->angle > PI * 2.0)
 		p->angle -= PI * 2.0;
-	p->dir_x = cos(p->angle);
-	p->dir_y = sin(p->angle);
+	p->dir_x = cosf(p->angle);
+	p->dir_y = sinf(p->angle);
 }
